requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-decouple
jinja2
python-multipart